test: should load the .env.test
