+++ jest.setup.ts
@@ -1,13 +1,4 @@
-import prettyFormat from 'pretty-format';
+import dotenv from 'dotenv';
+import path from 'path';
 
-global['prettyFormat'] = prettyFormat;
-
-jest.setTimeout(300000);
-
-// 把 console.error 转换成 error，方便断言
-(() => {
-  const spy = jest.spyOn(console, 'error');
-  afterAll(() => {
-    spy.mockRestore();
-  });
-})();
+dotenv.config({ path: path.resolve(process.cwd(), '.env.test') });
